split out storage component in bridge indexer into its own file
